refactor: move channels_v2/telegram_channel.py to channels/telegram_channel.py
