+++ apps/channels/tests/channels/senders/test_telegram_sender.py
@@ -2,7 +2,7 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from apps.channels.channels_v2.telegram_channel import TelegramSender
+from apps.channels.telegram_channel import TelegramSender
 
 
 @pytest.fixture()
@@ -17,14 +17,14 @@ def sender(telebot):
 
 class TestSendText:
     def test_sends_text_message(self, sender, telebot):
-        with patch("apps.channels.channels_v2.telegram_channel.antiflood") as mock_antiflood:
+        with patch("apps.channels.telegram_channel.antiflood") as mock_antiflood:
             sender.send_text("hello", "12345")
         mock_antiflood.assert_called_once_with(telebot.send_message, "12345", text="hello")
 
     def test_long_text_is_split_into_chunks(self, sender, telebot):
         # smart_split limits chunks to ~4096 chars; build a payload large enough to force splitting
         long_text = "a" * 8500
-        with patch("apps.channels.channels_v2.telegram_channel.antiflood") as mock_antiflood:
+        with patch("apps.channels.telegram_channel.antiflood") as mock_antiflood:
             sender.send_text(long_text, "12345")
         assert mock_antiflood.call_count >= 2
         for call in mock_antiflood.call_args_list:
@@ -34,7 +34,7 @@ class TestSendText:
             assert kwargs["text"]
 
     def test_propagates_send_errors(self, sender):
-        with patch("apps.channels.channels_v2.telegram_channel.antiflood", side_effect=RuntimeError("boom")):
+        with patch("apps.channels.telegram_channel.antiflood", side_effect=RuntimeError("boom")):
             with pytest.raises(RuntimeError, match="boom"):
                 sender.send_text("hello", "12345")
 
@@ -44,7 +44,7 @@ class TestSendVoice:
         audio = MagicMock()
         audio.audio = b"audio-bytes"
         audio.duration = 5
-        with patch("apps.channels.channels_v2.telegram_channel.antiflood") as mock_antiflood:
+        with patch("apps.channels.telegram_channel.antiflood") as mock_antiflood:
             sender.send_voice(audio, "12345")
         mock_antiflood.assert_called_once_with(telebot.send_voice, "12345", voice=audio.audio, duration=audio.duration)
 
@@ -67,7 +67,7 @@ class TestSendFile:
         file.content_type = content_type
         file.file = b"file-bytes"
 
-        with patch("apps.channels.channels_v2.telegram_channel.antiflood") as mock_antiflood:
+        with patch("apps.channels.telegram_channel.antiflood") as mock_antiflood:
             sender.send_file(file, "12345", session_id=42)
 
         expected_method = getattr(telebot, expected_method_attr)
